#29 restructured javascript read TBE
