+++ javascript/README.md
@@ -1,3 +1,11 @@
+## Installation
+
+Follow these steps to install the necessary dependencies for the project:
+
+```bash
+npm install
+```
+
 ## isolate_header
 
 (Add your content here)
@@ -16,7 +24,16 @@
 
 ## read_TBE
 
-(Add your content here)
+To start parsing sample TBE files into JavaScript native objects, you can use one of the following commands:
+
+```bash
+npm start
+```
+or
+
+```bash
+node src/functions/readTBE
+```
 
 ## strip_header
