+++ javascript/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "javascript_tbe",
+  "name": "javascript",
   "version": "1.0.0",
   "description": "",
-  "main": "parse_tbe.js",
+  "main": "index.js",
   "scripts": {
-    "start": "node parse_tbe",
+    "start": "node src/functions/readTBE",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
