--- javascript_tbe/README.md
@@ -1,20 +1,0 @@
-## Installation
-
-Follow these steps to install the necessary dependencies for the project:
-
-```bash
-npm install
-```
-
-## Usage
-
-To start parsing sample TBE files into JavaScript native objects, you can use one of the following commands:
-
-```bash
-npm start
-```
-or
-
-```bash
-node parse_tbe
-```
